requirements.txt
streamlit==1.38.0
requests>=2.32
python-dotenv>=1.0
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
textstat>=0.7
sentence-transformers>=3.0
pdfplumber>=0.11
pypdf>=5.0
rapidfuzz>=3.9
